MAJ
empeche la supression du SU :-D

+++ htdocs/php/AJAX-php/supprimeUser.php
@@ -6,12 +6,14 @@
     $pdo->query("SET NAMES utf8");
     $pdo->query("SET CHARACTER SET 'utf8'");
 
-
+    if($_GET['id']==1){
+        $codeHTML = 'Vous ne pouvez pas supprimer le compte administrateur.';
+    }else{
     //requet sql
         $sql = $pdo -> prepare ("DELETE FROM UTILISATEUR WHERE IdUser = :Id ;");
         $sql->execute(array(":Id" => $_GET['id']));
         $codeHTML='profil supprimé';
-
+    }
 
 $pdo=null;
